+++ src/pulumi/core-dtap/analytics/databricks/engineering_workspace.py
@@ -1,14 +1,9 @@
 from os import getenv
 
-import pulumi
 import pulumi_azure_native as azure_native
 import pulumi_azuread as azuread
 from pulumi import FileAsset, ResourceOptions
-from pulumi_databricks import (
-    Provider as DatabricksProvider,
-    ProviderArgs as DatabricksProviderArgs,
-    databricks,
-)
+from pulumi_databricks import Provider as DatabricksProvider, databricks
 
 from ingenii_azure_data_platform.databricks import create_cluster
 from ingenii_azure_data_platform.iam import (